[blockchain]修复UTC时间通过uint32存储位深不够的问题

+++ drm-blockchain/src/main/solidity/License.sol
@@ -8,8 +8,8 @@ contract License {
     string private rightName;
     address private owner;
     bool private available;
-    uint32 private issueTime;
-    uint32 private expireTime;
+    uint64 private issueTime;
+    uint64 private expireTime;
 
     //constructor
     constructor(string memory _rightName, uint32 _issueTime, uint32 _expireTime) {
